adapting plumbery to have reentrant plugin operations
Some functions are desirable to have with plugins, and even more, also
as helper functions available to eBPF programs. For instance, consider
the decode_frame operation. It is used in several places, and it is thus
interesting to make it available at numerous locations, especially with
possible extensions.

However, it needs to take arguments, which vary depending on the calling
context. The solution of putting all desired arguments as dedicated
fields in the context does not work anymore...

Instead, we rely on generic arguments stored in the context as uint64_t.
This is OK for pointers and for integers, which constitute the majority
(if not all) of types of arguments.

In addition, the plugin_run_operation has been modified to handle the
case of plugins calling plugins through external functions. It saves the
previous arguments and also cope with next operations that can possibly
be next ones of the caller plugin.

Notice that with the stress test, we do not observe any significant
overhead of this solution.