simplify the plumbery and be more powerful
So far, we could have a chain of protocol operations like
A -> B -> C -> B -> ...
but each state has to be independent of the running of the others, i.e.,
B cannot see the output arguments C has produced, it needed a (complex)
chain of identifiers to manage,...

Instead, when we plan to run a protocol operation, do it directly. It is
much simpler to develop code with this in mind, and allow the callee to
provide outputs to the caller with minimal overhead.