enable better opaque field sharing through API
This commit introduces the get_opaque_data field that, given an
identifier, returns the corresponding opaque field. This system is not
perfect, as it requires distinct opaque identifiers for each plugin.
Plugins thus still need to agree on different identifiers, but now they
can be in isolation with the opaque data of other possibly concurrent
plugins.