only one step away from seeing true multipath
We just now need three things. First, a path must be considered active
when a packet has been received on it. Second, we need to write in the
packet header the CID of the path on which the packet will be sent.
Third, let's send packets over multiple paths!

TODO: release path CIDs when deleting the context/paths.