change api to call protoop with possible parameter
We now introduce two types of protocol operations: those without
parameters, only defined by their ID, and the others with parameters
affecting their context. The idea is to allow plugins to hook the code
to, e.g., decode a particular frame, and the code with be run only for
that particular type frame, provided by the parameter.