performs packet computations when creating it
Do not perform a second pass on the packet to know if it requires
retransmission or not, for instance. This is a first needed step to
be able to detect if a packet needs to be retransmitted without
preparing the new packet itself.

This also removes the is_congestion_controlled protoop.