fix counting error of bytes_in_transit
When we send a packet, we count ALL the packet, with frames, checksum
overhead and packet header. When it got acknowledged, only the length of
frames and the checksum overhead is decreased. Therefore, the bytes in
flight values increases linearly (factor 13 for 13 bytes in the header)
with the number of packets sent. If a loss occurs, the CWIN decreases
below the bytes in transit value and the sender if definitely blocked.