MTU probes are now working
set_next_wake_time was very eager to send mtu probes but both the path and the
packet scheduler were not acting accordingly. Now MTU probes are no longer
congestion controlled, as recommended by DPLPMTUD. They are sent as soon as
required. A MTU probe is only required/sent if it's bigger than the current MTU.